migrate epacamd_eia subplant_id code into glue place & docs!
