+++ src/pudl/metadata/resources/glue.py
@@ -2,7 +2,7 @@
 from typing import Any
 
 RESOURCE_METADATA: dict[str, dict[str, Any]] = {
-    "epacamd_eia": {
+    "clean_epacamd_eia": {
         "description": """This crosswalk table comes from the
 EPA's Github repo camd-eia-crosswalk:
 https://github.com/USEPA/camd-eia-crosswalk.
@@ -17,8 +17,8 @@ have shifted before or since then aren't accurately reflected in the data. We're
 to create a temporal version of the crosswalk at some point.
 
 Our version of the crosswalk clarifies some of the column names and removes unmatched
-rows. The :func:`pudl.glue.epacamd_eia.transform` function doc strings explain what
-changes are made from the EPA's version.
+rows. The :func:`pudl.etl.glue_assets.clean_epacamd_eia` function doc strings explain
+what changes are made from the EPA's version.
 """,
         "schema": {
             "fields": [
@@ -40,11 +40,32 @@ crosswalk table which initally comes from the EPA's Github repo camd-eia-crosswa
 https://github.com/USEPA/camd-eia-crosswalk.
 It's purpose is to connect EPA units with EIA units, and generators.
 
+This table identifies subplants within plant_ids, which are the smallest coherent units
+for aggregation. plant_id refers to a legal entity that often contains multiple distinct
+power plants, even of different technology or fuel types.
+
+EPA CEMS data combines information from several parts of a power plant:
+* emissions from smokestacks
+* fuel use from combustors
+* electricty production from generators
+But smokestacks, combustors, and generators can be connected in complex, many-to-many
+relationships. This complexity makes attribution difficult for, as an example,
+allocating pollution to energy producers. Furthermore, heterogeneity within plant_ids
+make aggregation to the parent entity difficult or inappropriate.
+
 This table inherits from the EPA's crosswalk, the ID's from EPA CAMD
 hourly_emissions_epacems table itself, the boiler_generator_assn_eia860 table and the
 generators_eia860 table. While the epacamd_eia table is the core backbone of the table,
 EPA CAMD id's ensure there is complete coverage of EPA CAMD reporting units. The EIA 860
 table addition ensures there is also complete coverage of those units as well.
+
+For more information about the how this subplant_id is made, see the documentation for
+pudl.etl.glue_assets.make_subplant_ids and pudl.etl.glue_assets.update_subplant_ids
+from: https://catalystcoop-pudl.readthedocs.io/en/latest/autoapi/index.html
+
+But by analyzing the relationships between combustors and generators,
+as provided in the epacamd_eia crosswalk, we can identify distinct power plants.
+These are the smallest coherent units of aggregation.
 
 This table does not have primary keys because the primary keys would have been:
 plant_id_eia, generator_id, subplant_id and emissions_unit_id_epa, but there are some
@@ -60,10 +81,9 @@ successfully mapped to EIA generators.
                 "emissions_unit_id_epa",
                 "generator_id",
             ],
-            # "primary_key": ["plant_id_eia", "subplant_id", "generator_id"],
         },
         "field_namespace": "glue",
-        "etl_group": "epacems",
+        "etl_group": "glue",
         "sources": ["epacamd_eia"],
     },
 }
